+++ src/plugin.hpp
@@ -20,6 +20,7 @@ extern Model* modelInfix;
 extern Model* modelInfixMicro;
 extern Model* modelStrip;
 extern Model* modelStripCon4;
+extern Model* modelStripBlock;
 extern Model* modelEightFace;
 extern Model* modelEightFaceX2;
 extern Model* modelMidiCat;